Refactor models.py into a submodule (#2374)
* Refactor models.py into a submodule

* Add __all__ for each model file

* Expose baselayer models under `skyportal.models`

* Consistently import models from skyportal.models instead of baselayer
