+++ skyportal/model_util.py
@@ -1,5 +1,5 @@
 from social_tornado.models import TornadoStorage
-from skyportal.models import DBSession, ACL, Role, User, Group, Token
+from skyportal.models import DBSession, ACL, Role, User, Token, Group
 from skyportal.enum_types import LISTENER_CLASSES, sqla_enum_types
 from baselayer.app.env import load_env
