+++ skyportal/models/facility_transaction.py
@@ -1,0 +1,49 @@
+__all__ = ['FacilityTransaction']
+
+from datetime import datetime
+
+import sqlalchemy as sa
+from sqlalchemy.dialects import postgresql as psql
+from sqlalchemy.orm import relationship
+
+from baselayer.app.models import Base
+
+
+class FacilityTransaction(Base):
+
+    created_at = sa.Column(
+        sa.DateTime,
+        nullable=False,
+        default=datetime.utcnow,
+        index=True,
+        doc="UTC time this FacilityTransaction was created.",
+    )
+
+    request = sa.Column(psql.JSONB, doc='Serialized HTTP request.')
+    response = sa.Column(psql.JSONB, doc='Serialized HTTP response.')
+
+    followup_request_id = sa.Column(
+        sa.ForeignKey('followuprequests.id', ondelete='SET NULL'),
+        index=True,
+        nullable=True,
+        doc="The ID of the FollowupRequest this message pertains to.",
+    )
+
+    followup_request = relationship(
+        'FollowupRequest',
+        back_populates='transactions',
+        doc="The FollowupRequest this message pertains to.",
+    )
+
+    initiator_id = sa.Column(
+        sa.ForeignKey('users.id', ondelete='SET NULL'),
+        index=True,
+        nullable=True,
+        doc='The ID of the User who initiated the transaction.',
+    )
+    initiator = relationship(
+        'User',
+        back_populates='transactions',
+        doc='The User who initiated the transaction.',
+        foreign_keys="FacilityTransaction.initiator_id",
+    )
