+++ skyportal/models/gcn.py
@@ -1,0 +1,265 @@
+__all__ = ['GcnNotice', 'GcnTag', 'GcnEvent']
+
+import sqlalchemy as sa
+from sqlalchemy.orm import relationship
+from sqlalchemy.orm import deferred
+from sqlalchemy.ext.hybrid import hybrid_property
+
+import gcn
+import lxml
+
+from baselayer.app.models import Base, DBSession, AccessibleIfUserMatches
+
+
+class GcnNotice(Base):
+    """Records of ingested GCN notices"""
+
+    update = delete = AccessibleIfUserMatches('sent_by')
+
+    sent_by_id = sa.Column(
+        sa.ForeignKey('users.id', ondelete='CASCADE'),
+        nullable=False,
+        index=True,
+        doc="The ID of the User who created this GcnNotice.",
+    )
+
+    sent_by = relationship(
+        "User",
+        foreign_keys=sent_by_id,
+        back_populates="gcnnotices",
+        doc="The user that saved this GcnNotice",
+    )
+
+    ivorn = sa.Column(
+        sa.String, unique=True, index=True, doc='Unique identifier of VOEvent'
+    )
+
+    notice_type = sa.Column(
+        sa.Enum(gcn.NoticeType),
+        nullable=False,
+        doc='GCN Notice type',
+    )
+
+    stream = sa.Column(
+        sa.String, nullable=False, doc='Event stream or mission (i.e., "Fermi")'
+    )
+
+    date = sa.Column(sa.DateTime, nullable=False, doc='UTC message timestamp')
+
+    dateobs = sa.Column(
+        sa.ForeignKey('gcnevents.dateobs', ondelete="CASCADE"),
+        nullable=False,
+        doc='UTC event timestamp',
+    )
+
+    content = deferred(
+        sa.Column(sa.LargeBinary, nullable=False, doc='Raw VOEvent content')
+    )
+
+    def _get_property(self, property_name, value=None):
+        root = lxml.etree.fromstring(self.content)
+        path = ".//Param[@name='{}']".format(property_name)
+        elem = root.find(path)
+        value = float(elem.attrib.get('value', '')) * 100
+        return value
+
+    @property
+    def has_ns(self):
+        return self._get_property(property_name="HasNS")
+
+    @property
+    def has_remnant(self):
+        return self._get_property(property_name="HasRemnant")
+
+    @property
+    def far(self):
+        return self._get_property(property_name="FAR")
+
+    @property
+    def bns(self):
+        return self._get_property(property_name="BNS")
+
+    @property
+    def nsbh(self):
+        return self._get_property(property_name="NSBH")
+
+    @property
+    def bbh(self):
+        return self._get_property(property_name="BBH")
+
+    @property
+    def mass_gap(self):
+        return self._get_property(property_name="MassGap")
+
+    @property
+    def noise(self):
+        return self._get_property(property_name="Terrestrial")
+
+
+class GcnTag(Base):
+    """Store qualitative tags for events."""
+
+    update = delete = AccessibleIfUserMatches('sent_by')
+
+    sent_by_id = sa.Column(
+        sa.ForeignKey('users.id', ondelete='CASCADE'),
+        nullable=False,
+        index=True,
+        doc="The ID of the User who created this GcnTag.",
+    )
+
+    sent_by = relationship(
+        "User",
+        foreign_keys=sent_by_id,
+        back_populates="gcntags",
+        doc="The user that saved this GcnTag",
+    )
+
+    dateobs = sa.Column(
+        sa.ForeignKey('gcnevents.dateobs', ondelete="CASCADE"),
+        nullable=False,
+        index=True,
+    )
+
+    text = sa.Column(sa.Unicode, nullable=False)
+
+
+class GcnEvent(Base):
+    """Event information, including an event ID, mission, and time of the event."""
+
+    update = delete = AccessibleIfUserMatches('sent_by')
+
+    sent_by_id = sa.Column(
+        sa.ForeignKey('users.id', ondelete='CASCADE'),
+        nullable=False,
+        index=True,
+        doc="The ID of the User who created this GcnEvent.",
+    )
+
+    sent_by = relationship(
+        "User",
+        foreign_keys=sent_by_id,
+        back_populates="gcnevents",
+        doc="The user that saved this GcnEvent",
+    )
+
+    dateobs = sa.Column(sa.DateTime, doc='Event time', unique=True, nullable=False)
+
+    gcn_notices = relationship("GcnNotice", order_by=GcnNotice.date)
+
+    _tags = relationship(
+        "GcnTag",
+        order_by=(
+            sa.func.lower(GcnTag.text).notin_({'fermi', 'swift', 'amon', 'lvc'}),
+            sa.func.lower(GcnTag.text).notin_({'long', 'short'}),
+            sa.func.lower(GcnTag.text).notin_({'grb', 'gw', 'transient'}),
+        ),
+    )
+
+    localizations = relationship("Localization")
+
+    @hybrid_property
+    def tags(self):
+        """List of tags."""
+        return [tag.text for tag in self._tags]
+
+    @tags.expression
+    def tags(cls):
+        """List of tags."""
+        return (
+            DBSession()
+            .query(GcnTag.text)
+            .filter(GcnTag.dateobs == cls.dateobs)
+            .subquery()
+        )
+
+    @hybrid_property
+    def retracted(self):
+        """Check if event is retracted."""
+        return 'retracted' in self.tags
+
+    @retracted.expression
+    def retracted(cls):
+        """Check if event is retracted."""
+        return sa.literal('retracted').in_(cls.tags)
+
+    @property
+    def lightcurve(self):
+        """GRB lightcurve URL."""
+        try:
+            notice = self.gcn_notices[0]
+        except IndexError:
+            return None
+        root = lxml.etree.fromstring(notice.content)
+        elem = root.find(".//Param[@name='LightCurve_URL']")
+        if elem is None:
+            return None
+        else:
+            try:
+                return elem.attrib.get('value', '').replace('http://', 'https://')
+            except Exception:
+                return None
+
+    @property
+    def gracesa(self):
+        """Event page URL."""
+        try:
+            notice = self.gcn_notices[0]
+        except IndexError:
+            return None
+        root = lxml.etree.fromstring(notice.content)
+        elem = root.find(".//Param[@name='EventPage']")
+        if elem is None:
+            return None
+        else:
+            try:
+                return elem.attrib.get('value', '')
+            except Exception:
+                return None
+
+    @property
+    def ned_gwf(self):
+        """NED URL."""
+        return "https://ned.ipac.caltech.edu/gwf/events"
+
+    @property
+    def HasNS(self):
+        """Checking if GW event contains NS."""
+        notice = self.gcn_notices[0]
+        root = lxml.etree.fromstring(notice.content)
+        elem = root.find(".//Param[@name='HasNS']")
+        if elem is None:
+            return None
+        else:
+            try:
+                return 'HasNS: ' + elem.attrib.get('value', '')
+            except Exception:
+                return None
+
+    @property
+    def HasRemnant(self):
+        """Checking if GW event has remnant matter."""
+        notice = self.gcn_notices[0]
+        root = lxml.etree.fromstring(notice.content)
+        elem = root.find(".//Param[@name='HasRemnant']")
+        if elem is None:
+            return None
+        else:
+            try:
+                return 'HasRemnant: ' + elem.attrib.get('value', '')
+            except Exception:
+                return None
+
+    @property
+    def FAR(self):
+        """Returning event false alarm rate."""
+        notice = self.gcn_notices[0]
+        root = lxml.etree.fromstring(notice.content)
+        elem = root.find(".//Param[@name='FAR']")
+        if elem is None:
+            return None
+        else:
+            try:
+                return 'FAR: ' + elem.attrib.get('value', '')
+            except Exception:
+                return None
