+++ skyportal/tests/fixtures.py
@@ -15,6 +15,7 @@ from baselayer.app.config import load_config
 from baselayer.app.env import load_env
 from baselayer.app.test_util import set_server_url
 from skyportal.models import (
+    init_db,
     DBSession,
     User,
     Group,
@@ -32,7 +33,6 @@ from skyportal.models import (
     ClassicalAssignment,
     Taxonomy,
     Classification,
-    init_db,
     FollowupRequest,
     Allocation,
     Invitation,
